+++ guides.ts
@@ -7,6 +7,10 @@ export default [
     },
     {
         name: 'DesignGuidelines',
-        children: ['src/color-system.md', 'src/size-rhythms.md'],
+        children: [
+            'src/color-system.md',
+            'src/decluttering.md',
+            'src/size-rhythms.md',
+        ],
     },
 ] as Guide[];