+++ src/components/collapsible-section/collapsible-section.tsx
@@ -18,7 +18,6 @@ import { Action } from './action';
 export class CollapsibleSection {
     /**
      * `true` if the section is expanded, `false` if collapsed.
-     * Defaults to `false`.
      */
     @Prop({ mutable: true, reflect: true })
     public isOpen: boolean = false;
@@ -61,28 +60,19 @@ export class CollapsibleSection {
     public render() {
         return (
             <section class={`${this.isOpen ? 'open' : ''}`}>
-                <header onClick={this.onClick}>
-                    <span class="expander">
-                        <svg viewBox="0 0 30 30" width="100%" height="100%">
-                            {(() => {
-                                if (this.isOpen) {
-                                    return (
-                                        <path d="M15.5 12.246l1.008 1.008L21 17.746l-1.008 1.008-4.492-4.491-4.492 4.492L10 17.746l4.492-4.492 1.008-1.008z" />
-                                    );
-                                }
-
-                                return (
-                                    <path d="M15.5 19l1.008-1.008L21 13.5l-1.008-1.008-4.492 4.491-4.492-4.492L10 13.5l4.492 4.492L15.5 19z" />
-                                );
-                            })()}
-                        </svg>
-                    </span>
-                    <h2 class="mdc-typography mdc-typography--headline2">
+                <header class="section__header" onClick={this.onClick}>
+                    <div class="section__header__expand-icon">
+                        <div class="expand-icon__line"></div>
+                        <div class="expand-icon__line"></div>
+                        <div class="expand-icon__line"></div>
+                        <div class="expand-icon__line"></div>
+                    </div>
+                    <h2 class="section__header__title mdc-typography mdc-typography--headline2">
                         {this.header}
                     </h2>
                     {this.renderActions()}
                 </header>
-                <div class="body">
+                <div class="section__body">
                     <slot />
                 </div>
             </section>
@@ -107,7 +97,7 @@ export class CollapsibleSection {
         }
 
         return (
-            <div class="actions">
+            <div class="section__header__actions">
                 {this.actions.map(this.renderActionButton)}
             </div>
         );