+++ src/components/tab-panel/tab-panel.tsx
@@ -6,25 +6,26 @@ import {
     EventEmitter,
     Event,
     Host,
+    Watch,
 } from '@stencil/core';
 import { Tab } from '../tab-bar/tab.types';
 
 /**
- * The `tab-panel` component uses the `tab-bar` component together with custom
- * slotted components and will display the content for the currently active tab.
- * Each slotted component must have an id equal to the id of the corresponding
- * tab it belongs to.
+ * The `limel-tab-panel` component uses the `limel-tab-bar` component together
+ * with custom slotted components and will display the content for the currently
+ * active tab. Each slotted component must have an id equal to the id of the
+ * corresponding tab it belongs to.
  *
- * The `tab-panel` component will automatically set each tab configuration on
- * the corresponding slotted component as a property named `tab` so that the
+ * The `limel-tab-panel` component will automatically set each tab configuration
+ * on the corresponding slotted component as a property named `tab` so that the
  * component can take action upon that. Sometimes it might be desirable to not
  * load data or render anything until the tab is active.
  *
  * The slotted components can also emit the `changeTab` event to update anything
  * inside the actual tab, e.g. to change the icon, color or badge.
  *
- * @slot - Content to put inside the `tab-panel`. Each slotted element must
- * have the `id` attribute equal to the id of the tab it belongs to.
+ * @slot - Content to put inside the `limel-tab-panel`. Each slotted element
+ * must have the `id` attribute equal to the id of the tab it belongs to.
  *
  * @exampleComponent limel-example-tab-panel
  */
@@ -81,6 +82,12 @@ export class TabPanel {
         slot.removeEventListener('slotchange', this.setSlotElements);
     }
 
+    @Watch('tabs')
+    public tabsChanged() {
+        this.hidePanels();
+        this.tabs.forEach(this.setTabStatus);
+    }
+
     public render() {
         return (
             <Host onChangeTab={this.handleChangeTabs}>
@@ -96,7 +103,9 @@ export class TabPanel {
 
     private setSlotElements() {
         const slot = this.getSlot();
+        this.hidePanels();
         this.slotElements = [].slice.call(slot.assignedElements());
+        this.tabs.forEach(this.setTabStatus);
     }
 
     private setTabStatus(tab: Tab) {
@@ -128,5 +137,11 @@ export class TabPanel {
 
     private getSlot(): HTMLSlotElement {
         return this.host.shadowRoot.querySelector('slot');
+    }
+
+    private hidePanels() {
+        this.slotElements.forEach((element) => {
+            element.style.display = 'none';
+        });
     }
 }